pyproject.toml
[project]
name = "jenkins-job-insight"
version = "3.2.0"
description = "Jenkins job insight and analysis tool"
requires-python = ">=3.12"
dependencies = [
  "ai-cli-runner>=0.1.1",
  "aiosqlite",
  "cryptography>=46.0.5",
  "defusedxml>=0.7.1",
  "fastapi",
  "gitpython",
  "httpx",
  "pydantic-settings",
  "python-jenkins",
  "python-multipart",
  "pyyaml",
  "pywebpush",
  "python-simple-logger",
  "reportportal-client>=5.7.4",
  "typer>=0.9.0",
  "uvicorn",
  "pyyaml>=6.0.3",
]

[project.scripts]
jenkins-job-insight = "jenkins_job_insight.main:run"
jji = "jenkins_job_insight.cli.main:app"

[project.optional-dependencies]
tests = [
    "pytest",
    "pytest-asyncio",
    "pytest-xdist>=3.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["src"]

[tool.hatch.build.targets.wheel]
packages = ["src/jenkins_job_insight"]
